yacmode() history (except possibly Mac)

+++ R/yacmode.R
@@ -5,9 +5,11 @@
 #' The user types valid yacas input and presses return. Type 'quit' to return
 #' to R prompt.
 #' 
+#' 
+#' @param enable_history Use `R` history such that previous yacas commands can be used. Default is `TRUE`.
 #' @return Output of yacas is returned.
-#' @note Note that command recall will recall previous R commands, not previous
-#' yacas input. Yacas is given a limited amount of time to complete, otherwise
+#' @note Note that command will use R `history()` and modify it by default. 
+#' Yacas is given a limited amount of time to complete, otherwise
 #' `\[1\] CommandLine(1) : User interrupted calculation` is returned. E.g.
 #' Taylor(x,0,5) 1/(1+x) will work, but Taylor(x,0,12) 1/(1+x) is likely to
 #' take too long.
@@ -23,27 +25,23 @@
 #' }
 #' 
 #' @export
-yacmode <-function () {
+yacmode <- function(enable_history = TRUE) {
   # Enable history in yacmode()
   # https://stackoverflow.com/a/27528113
-  tmphistory <- tempfile()
-  savehistory(tmphistory)
-  on.exit(unlink(tmphistory))
+  if (enable_history == TRUE) {
+    tmphistory <- tempfile()
+    savehistory(tmphistory)
+    on.exit(unlink(tmphistory))
+  }
   
   cat("Enter Yacas commands here. Type quit to return to R\n")
   x <- readline("Yacas->")
   while (length(which(c("stop;", "stop", "end;", "end", "quit;",
-                        "quit", "exit;", "exit", "e;", "e", "q;", "q","\n") == x)) ==
+                        "quit", "exit;", "exit", "e;", "e", "q;", "q", "q()", "\n") == x)) ==
          0) {
 
     o <- yacas(x)
     print(o)
-    
-    # Update history:
-    histcon <- file(tmphistory, open = "a")
-    writeLines(x, histcon)
-    close(histcon)
-    loadhistory(tmphistory)
     
     # Old:
     # x <- gsub("Out>","#",x)
@@ -63,6 +61,14 @@ yacmode <-function () {
     # }
     
     x <- readline("Yacas->")
+
+    # Update history:
+    if (enable_history == TRUE) {
+      histcon <- file(tmphistory, open = "a")
+      writeLines(x, histcon)
+      close(histcon)
+      loadhistory(tmphistory)
+    }
   }
 }
